Dearie me.  This checkin actually works, although I still haven't implemented the output.c patch needed by Dirk or fixed the missing .h file in include/dieharder.  Still, I >>HAVE<< turned all the tests into type int's (still no returns) and stripped dieharder.h and split off a globals.h file that I don't think I'm going to need, actually, although it was useful while stripping dieharder.h as a reservoir of codelets that I needed to put back.
Anyway, it is entirely possible that Dirk will read these words as I
ALSO have dieharder.googlecode.com set up with him on it, and while this
checkin is still local (about to be svnsync'd up, not directly checked
in) VERY SOON NOW I may try checkout out from the google repo, which
will of course check BACK into the google repo thereafter.

+++ libdieharder/rgb_lmn.c
@@ -93,7 +93,7 @@
 #include <dieharder/libdieharder.h>
 #include <dieharder/rgb_lmn.h>
 
-double rgb_lmn()
+int rgb_lmn()
 {
 
  double pks;
